Fix integer types & struct size in zlib.nim
Integer types in zlib.h were not properly translated into Nim. Most
seriously, `Ulong` and `Ulongf` (which are both `unsigned long`)
were translated to `int32`, which is wrong on LP64 systems like
macOS, iOS and I think 64-bit Linux too.

This causes the Nim `ZStream` struct to be much shorter than the real
`z_stream` (88 vs 120 bytes), and that causes the size comparison in
the inflate/deflateInit functions to fail with Z_VERSION_ERROR, which
is actually a good thing because if it kept going it would be smashing
memory, since Nim didn't allocate enough space for its struct.

I've fixed the types to the best of my knowledge, using the standard
C-based types like `cint` and `culong`. The only iffy one is `ZOffT`
which has a complicated conditional definition in C -- I think it's
supposed to match `off_t`, whose size depends not on the ABI but on
whether the filesystem supports 64-bit files... I made it a `clong`
since most systems handle 64-bit files nowadays.

Fixes #23, #39, #53